@@ -4,22 +4,18 @@ import (
 	"bytes"
 	"fmt"
 	"go/ast"
-	"go/build"
 	"go/format"
-	"go/parser"
 	"go/printer"
 	"go/token"
 	"go/types"
-	"io/ioutil"
 	"log"
 	"os"
 	"path"
 	"sort"
 	"strings"
-	"unicode"
 
 	"github.com/google/go-cmp/cmp"
-	"golang.org/x/tools/go/loader"
+	"golang.org/x/tools/go/packages"
 )
 
 const (
@@ -28,20 +24,30 @@ const (
 
 // GenerateBundleFile generates a file with static definitions for Gorums.
 func GenerateBundleFile(dst string) {
-	pkgIdent, reservedIdents, code := bundle(devPkgPath)
-	// escape backticks
-	escaped := strings.ReplaceAll(string(code), "`", "`+\"`\"+`")
-	src := "// Code generated by protoc-gen-gorums. DO NOT EDIT.\n" +
-		"// Source files can be found in: " + devPkgPath + "\n\n" +
-		"package gengorums\n\n" +
-		generatePkgMap(pkgIdent, reservedIdents) +
-		"var staticCode = `" + escaped + "`\n"
+	pkg := loadPackage(devPkgPath)
+	code := printFiles(pkg)
+	pkgIdentMap, reservedIdents := findIdentifiers(pkg)
+
+	src := fmt.Sprintf(`// Code generated by protoc-gen-gorums. DO NOT EDIT.
+	// Source files can be found in: %s
+
+	package gengorums
+
+	// pkgIdentMap maps from package name to one of the package's identifiers.
+	// These identifiers are used by the Gorums protoc plugin to generate import statements.
+	var pkgIdentMap = %#v
+
+	// reservedIdents holds the set of Gorums reserved identifiers.
+	// These identifiers cannot be used to define message types in a proto file.
+	var reservedIdents = %#v
+
+	var staticCode = `+"`%s`", devPkgPath, pkgIdentMap, reservedIdents, code)
 
 	staticContent, err := format.Source([]byte(src))
 	if err != nil {
 		log.Fatalf("formatting failed: %v", err)
 	}
-	currentContent, err := ioutil.ReadFile(dst)
+	currentContent, err := os.ReadFile(dst)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -50,62 +56,37 @@ func GenerateBundleFile(dst string) {
 		fmt.Fprintf(os.Stderr, "\nReview changes above; to revert use:\n")
 		fmt.Fprintf(os.Stderr, "mv %s.bak %s\n", dst, dst)
 	}
-	err = ioutil.WriteFile(dst, []byte(staticContent), 0666)
+	err = os.WriteFile(dst, []byte(staticContent), 0666)
 	if err != nil {
 		log.Fatal(err)
 	}
 }
 
-func generatePkgMap(pkgs map[string]string, reservedIdents []string) string {
-	var keys []string
-	for k := range pkgs {
-		keys = append(keys, k)
+// loadPackage returns the parsed package.
+func loadPackage(pkgPath string) *packages.Package {
+	cfg := &packages.Config{
+		Mode: packages.NeedName | packages.NeedFiles | packages.NeedSyntax | packages.NeedImports | packages.NeedTypes | packages.NeedTypesInfo,
 	}
-	sort.Strings(keys)
-	var buf bytes.Buffer
-	buf.WriteString(`
-	// pkgIdentMap maps from package name to one of the package's identifiers.
-	// These identifiers are used by the Gorums protoc plugin to generate
-	// appropriate import statements.
-	`)
-	buf.WriteString("var pkgIdentMap = map[string]string{\n")
-	for _, imp := range keys {
-		buf.WriteString("\t" + `"`)
-		buf.WriteString(imp)
-		buf.WriteString(`": "`)
-		buf.WriteString(pkgs[imp])
-		buf.WriteString(`",` + "\n")
+	pkgs, err := packages.Load(cfg, pkgPath)
+	if err != nil {
+		log.Fatalf("failed to load %s: %v", pkgPath, err)
 	}
-	buf.WriteString("}\n\n")
-	buf.WriteString(`
-	// reservedIdents holds the set of Gorums reserved identifiers.
-	// These identifiers cannot be used to define message types in a proto file.
-	`)
-	buf.WriteString("var reservedIdents = []string{\n")
-	for _, ident := range reservedIdents {
-		buf.WriteString("\t" + `"`)
-		buf.WriteString(ident)
-		buf.WriteString(`",` + "\n")
+	if packages.PrintErrors(pkgs) > 0 {
+		os.Exit(1)
 	}
-	buf.WriteString("}\n\n")
-	return buf.String()
+	// Since Load succeeded and pkgPath is a single package, the following is safe
+	return pkgs[0]
 }
 
-// findIdentifiers examines the given package to find all imported packages,
-// and one used identifier in that imported package. These identifiers are
-// used by the Gorums protoc plugin to generate appropriate import statements.
-func findIdentifiers(fset *token.FileSet, info *loader.PackageInfo) (map[string]string, []string) {
-	packagePath := info.Pkg.Path()
+// findIdentifiers returns the imported packages as a map from package name to one of its identifiers,
+// and a slice of Gorums reserved identifiers that cannot be used in proto files.
+// The returned map is used by the Gorums protoc plugin to generate import statements.
+func findIdentifiers(pkgInfo *packages.Package) (map[string]string, []string) {
 	pkgIdents := make(map[string][]string)
-	for id, obj := range info.Uses {
-		pos := fset.Position(id.Pos())
-		if strings.Contains(pos.Filename, "zorums") {
-			// ignore identifiers in zorums generated files
-			continue
-		}
-		r := []rune(obj.Name())
-		if unicode.IsLower(r[0]) {
-			// ignore unexported identifiers
+	for id, obj := range pkgInfo.TypesInfo.Uses {
+		pos := pkgInfo.Fset.Position(id.Pos())
+		if ignore(pos.Filename) || !obj.Exported() {
+			// ignore identifiers in zorums generated files and unexported identifiers
 			continue
 		}
 		if pkg := obj.Pkg(); pkg != nil {
@@ -138,7 +119,7 @@ func findIdentifiers(fset *token.FileSet, info *loader.PackageInfo) (map[string]
 	var reservedIdents []string
 	for path, idents := range pkgIdents {
 		sort.Strings(idents)
-		if path == packagePath {
+		if path == pkgInfo.PkgPath {
 			reservedIdents = idents
 			continue
 		}
@@ -157,31 +138,12 @@ func addUniqueIdentifier(pkgIdents map[string][]string, path, name string) {
 	pkgIdents[path] = append(pkgIdents[path], name)
 }
 
-// bundle returns a slice with the code for the given src package without imports.
-// The returned map contains packages to be imported along with one identifier
-// using the relevant import path.
-// Loosly based on x/tools/cmd/bundle
-func bundle(src string) (map[string]string, []string, []byte) {
-	conf := loader.Config{ParserMode: parser.ParseComments, Build: &build.Default}
-	conf.Import(src)
-	lprog, err := conf.Load()
-	if err != nil {
-		log.Fatalf("failed to load Go package: %v", err)
-	}
-	// Because there was a single Import call and Load succeeded,
-	// InitialPackages is guaranteed to hold the sole requested package.
-	info := lprog.InitialPackages()[0]
-
-	var out bytes.Buffer
-	printFiles(&out, lprog.Fset, info)
-	pkgIdentMap, reservedIdents := findIdentifiers(lprog.Fset, info)
-	return pkgIdentMap, reservedIdents, out.Bytes()
-}
-
-func printFiles(out *bytes.Buffer, fset *token.FileSet, info *loader.PackageInfo) {
-	for _, f := range info.Files {
+// printFiles returns the source files of pkg without package and import statements.
+func printFiles(pkg *packages.Package) string {
+	out := &bytes.Buffer{}
+	for _, f := range pkg.Syntax {
 		// filter files in dev package that shouldn't be bundled in template_static.go
-		fileName := fset.File(f.Pos()).Name()
+		fileName := pkg.Fset.File(f.Pos()).Name()
 		if ignore(fileName) {
 			continue
 		}
@@ -208,16 +170,18 @@ func printFiles(out *bytes.Buffer, fset *token.FileSet, info *loader.PackageInfo
 			printComments(out, f.Comments, last, beg)
 
 			buf.Reset()
-			err := format.Node(&buf, fset, &printer.CommentedNode{Node: decl, Comments: f.Comments})
+			err := format.Node(&buf, pkg.Fset, &printer.CommentedNode{Node: decl, Comments: f.Comments})
 			if err != nil {
 				log.Fatalf("failed to format source AST node: %v", err)
 			}
 			out.Write(buf.Bytes())
-			last = printSameLineComment(out, f.Comments, fset, end)
+			last = printSameLineComment(out, f.Comments, pkg.Fset, end)
 			out.WriteString("\n\n")
 		}
 		printLastComments(out, f.Comments, last)
 	}
+	// escape backticks
+	return strings.ReplaceAll(out.String(), "`", "`+\"`\"+`")
 }
 
 // ignore files in dev folder with suffixes that shouldn't be bundled.
