@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
 // 	protoc-gen-gorums v0.3.0-devel
-// 	protoc            v3.15.5
+// 	protoc            v3.15.6
 // source: storage/proto/storage.proto
 
 package proto
