@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
 // 	protoc-gen-go v1.25.0
-// 	protoc        v3.15.5
+// 	protoc        v3.15.6
 // source: internal/correctable/opts.proto
 
 package correctable
